+++ src/server.rs
@@ -71,7 +71,7 @@ pub struct ServerPlugin {
     /// If disabled, replication should be started manually by inserting [`ReplicatedClient`] on the client entity.
     ///
     /// Until replication has started, the client and server can still exchange network events that are marked as
-    /// independent via [`ServerEventAppExt::make_independent`](crate::shared::event::server_event::ServerEventAppExt::make_independent).
+    /// independent via [`ServerEventAppExt::make_event_independent`](crate::shared::event::server_event::ServerEventAppExt::make_event_independent).
     /// **All other events will be ignored**.
     pub replicate_after_connect: bool,
 }